customer-portal/backend/pyproject.toml
[project]
name = "customer-portal-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "python-multipart>=0.0.9",
    "python-magic>=0.4.27",
    "clamd>=1.0.2",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
]
